move nonemptyarray to utils
