--- packages/repocop/src/query.test.ts
@@ -1,15 +1,0 @@
-import { toNonEmptyArray } from './query';
-import type { NonEmptyArray } from './query';
-
-describe('Failure on empty arrays', () => {
-	test('throw an error if input is an empty array', () => {
-		const emptyArray: string[] = [];
-		const nonEmptyArray: string[] = ['a', 'b'];
-		const typedNonEmptyArray: NonEmptyArray<string> = ['a', 'b'];
-
-		expect(() => toNonEmptyArray(emptyArray)).toThrow();
-		expect(() => toNonEmptyArray(nonEmptyArray)).not.toThrow();
-		expect(toNonEmptyArray(nonEmptyArray)).toEqual(nonEmptyArray);
-		expect(toNonEmptyArray(nonEmptyArray)).toEqual(typedNonEmptyArray);
-	});
-});
