+++ packages/repocop/src/utils.test.ts
@@ -1,5 +1,5 @@
-import type { Repository } from './types';
-import { isProduction } from './utils';
+import type { NonEmptyArray, Repository } from './types';
+import { isProduction, toNonEmptyArray } from './utils';
 
 describe('isProduction', () => {
 	test('should return correct values for prod and non-prod repos', () => {
@@ -21,5 +21,18 @@ describe('isProduction', () => {
 
 		expect(isProduction(prodRepo)).toBe(true);
 		expect(isProduction(nonProdRepo)).toBe(false);
+	});
+});
+
+describe('Failure on empty arrays', () => {
+	test('throw an error if input is an empty array', () => {
+		const emptyArray: string[] = [];
+		const nonEmptyArray: string[] = ['a', 'b'];
+		const typedNonEmptyArray: NonEmptyArray<string> = ['a', 'b'];
+
+		expect(() => toNonEmptyArray(emptyArray)).toThrow();
+		expect(() => toNonEmptyArray(nonEmptyArray)).not.toThrow();
+		expect(toNonEmptyArray(nonEmptyArray)).toEqual(nonEmptyArray);
+		expect(toNonEmptyArray(nonEmptyArray)).toEqual(typedNonEmptyArray);
 	});
 });
