+++ packages/repocop/src/utils.ts
@@ -1,7 +1,14 @@
-import type { Repository } from './types';
+import type { NonEmptyArray, Repository } from './types';
 
 export function isProduction(repo: Repository) {
 	return repo.topics.includes('production') && !repo.archived;
+}
+
+export function toNonEmptyArray<T>(value: T[]): NonEmptyArray<T> {
+	if (value.length === 0) {
+		throw new Error(`Expected a non-empty array. Source table may be empty.`);
+	}
+	return value as NonEmptyArray<T>;
 }
 
 //TODO test this
